Test runs on Maven Surefire now
Technically, it is still an integration test, but as it no longer needs the JAR
from the 'package' phase, it can run via Surefire and I can get rid of the Failsafe
configuration.

I also removed the redundant Maven Compiler declaration because the plugin is
actually not needed in the latest version.